Add files via upload
These are the files in that I was trying to create a tfrecords file for all the data so that it would be easier to train with and augment the data. Had an issue with trying to encode the images as pngs files otherwise was able to encode and decode them too. The tfrecords files ive uploaded to the one drive; there are three; one dataset with the angles as targets, one with the speed and then one with both as targets. Also I treated them as classes rather than contiunous values since there are only 17 angles and 2 speeds in the entire 2000 examples I looked through. The issue with being encoded as something other than pngs in the records files comes from the functions 'convert_1_label' and 'convert_2_label' in creating_files.py', where im.tostring() encodes them as something other than pngs.  The reading_files.py is then supposed to decode the records and pass them onto the network to train. The idea was to then do all the data augmenting on the fly in the 'training_input_function' and 'training_input_function' functions in the files 'reading_files.py' rather than manually. If you're feeling brave, please have a look, its a mess but should make some sense.